+++ src/Fragment.js
@@ -1,0 +1,4 @@
+import {controlFlow} from "./controlFlow.js";
+import {fragment} from "./mount.js";
+
+export const Fragment = controlFlow(fragment);
