+++ src/method.js
@@ -1,0 +1,5 @@
+export const method =
+  (fn) =>
+  (d) =>
+  (...params) =>
+    fn(d, ...params);
