+++ src/ref.js
@@ -1,0 +1,5 @@
+import {symbol} from "./shared.js";
+
+export const setRef = symbol();
+
+export const ref = symbol();
